pyproject.toml
[project]
name = "flaskproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "jinja2>=3.1.6",
    "mypy>=1.15.0",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.0",
    "python-slugify>=8.0.4",
    "ruff>=0.11.10",
    "types-flask-migrate>=4.1.0.20250112",
    "validators>=0.35.0",
]
